+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.leocardz.link.preview"
+    package="com.leocardz.linkpreview.sample"
     android:installLocation="preferExternal">
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -12,7 +12,7 @@
         android:screenOrientation="portrait"
         android:theme="@style/App.Theme">
         <activity
-            android:name=".Main"
+            android:name="com.leocardz.linkpreview.sample.Main"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             android:theme="@style/App.Theme">
