--- library/build.gradle
@@ -1,38 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android-extensions'
-apply plugin: 'kotlin-android'
-
-android {
-    compileSdkVersion 29
-
-    defaultConfig {
-        minSdkVersion 9
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.3.0"
-
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    implementation 'org.jsoup:jsoup:1.11.3'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.12'
-
-    androidTestImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'org.mockito:mockito-android:2.22.0'
-}
-
-repositories {
-    mavenCentral()
-}
