--- library/src/main/AndroidManifest.xml
@@ -1,5 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.leocardz.link.preview.library">
-
-    <uses-permission android:name="android.permission.INTERNET" />
-</manifest>
