+++ linkpreview/proguard-rules.pro
