+++ linkpreview/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.leocardz.linkpreview.sample.library">
+
+    <uses-permission android:name="android.permission.INTERNET" />
+</manifest>
