+++ linkpreview/src/main/java/com/leocardz/linkpreview/sample/library/DowloadImageCallback.java
@@ -1,0 +1,19 @@
+package com.leocardz.linkpreview.sample.library;
+
+import android.graphics.Bitmap;
+import android.widget.ImageView;
+
+public interface DowloadImageCallback {
+
+	/**
+	 * 
+	 * @param imageView
+	 *            ImageView to receive the bitmap.
+	 * @param loadedBitmap
+	 *            Bitmap downloaded from url.
+	 * @param url
+	 *            Image url.
+	 */
+	void onLoaded(ImageView imageView, Bitmap loadedBitmap, String url);
+
+}
