+++ linkpreview/src/main/java/com/leocardz/linkpreview/sample/library/GetSource.kt
@@ -1,0 +1,15 @@
+package com.leocardz.linkpreview.sample.library
+
+/**
+ * @author dhruvaraj nagarajan
+ */
+interface GetSource {
+
+    fun getSourceCode(url: String): SourceContent
+
+    fun isNull(sourceContent: SourceContent): Boolean
+
+    fun cancel()
+
+    fun isCallCancelled(): Boolean
+}
