+++ linkpreview/src/main/java/com/leocardz/linkpreview/sample/library/ImagePickingStrategy.kt
@@ -1,0 +1,14 @@
+package com.leocardz.linkpreview.sample.library
+
+import org.jsoup.nodes.Document
+import java.util.*
+
+/**
+ * A strategy for how to select the images to return.
+ */
+interface ImagePickingStrategy {
+
+    var imageQuantity: Int
+
+    fun getImages(getSource: GetSource, doc: Document, metaTags: HashMap<String, String>): List<String>
+}
