+++ linkpreview/src/main/java/com/leocardz/linkpreview/sample/library/url/UrlExtractionStrategy.java
@@ -1,0 +1,11 @@
+package com.leocardz.linkpreview.sample.library.url;
+
+import java.util.List;
+
+/**
+ * Provides the means for extracting URL(s) from text.
+ */
+public interface UrlExtractionStrategy {
+
+    List<String> extractUrls(String textPassedToTextCrawler);
+}
